Migrate `unit.serializers` to PyTest
